+++ src/main/java/dinkplugin/notifiers/DeathNotifier.java
@@ -15,7 +15,6 @@ import dinkplugin.util.ItemUtils;
 import dinkplugin.util.Region;
 import dinkplugin.util.Utils;
 import dinkplugin.util.WorldUtils;
-import lombok.Synchronized;
 import lombok.extern.slf4j.Slf4j;
 import net.runelite.api.Actor;
 import net.runelite.api.Item;
@@ -30,6 +29,7 @@ import net.runelite.api.Varbits;
 import net.runelite.api.events.ActorDeath;
 import net.runelite.api.events.InteractingChanged;
 import net.runelite.api.events.ScriptPreFired;
+import net.runelite.client.callback.ClientThread;
 import net.runelite.client.game.ItemManager;
 import net.runelite.client.game.NPCManager;
 import org.apache.commons.lang3.ArrayUtils;
@@ -104,6 +104,9 @@ public class DeathNotifier extends BaseNotifier {
     @Inject
     private NPCManager npcManager;
 
+    @Inject
+    private ClientThread clientThread;
+
     /**
      * Tracks the last {@link Actor} our local player interacted with,
      * for the purposes of attributing deaths to particular {@link Player}'s.
@@ -127,16 +130,16 @@ public class DeathNotifier extends BaseNotifier {
     }
 
     public void init() {
-        setIgnoredRegions(config.deathIgnoredRegions());
+        clientThread.invoke(() -> setIgnoredRegions(config.deathIgnoredRegions()));
     }
 
     public void reset() {
-        setIgnoredRegions(null);
+        clientThread.invoke(() -> setIgnoredRegions(null));
     }
 
     public void onConfigChanged(String key, String value) {
         if ("deathIgnoredRegions".equals(key)) {
-            setIgnoredRegions(value);
+            clientThread.invoke(() -> setIgnoredRegions(value));
         }
     }
 
@@ -299,7 +302,6 @@ public class DeathNotifier extends BaseNotifier {
         return isEnabled();
     }
 
-    @Synchronized
     private void setIgnoredRegions(@Nullable String configValue) {
         ignoredRegions.clear();
         ConfigUtil.readDelimited(configValue).forEach(str -> {