+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Minor: Allow customization of rich embed highlight color. (#534)
 - Minor: Include Moxi pet name in pet notifications. (#557)
 
 ## 1.10.10